Organize button module into directory structure
